// src/services/models/Metadata.ts
export interface Metadata {
    originAddress: string;
    originNeighborhood: string;
    originCity: string,
    originState: string;
    originCep: string;
    originCountry: string;
    destinationAddress: string;
    destinationNeighborhood: string;
    destinationCity: string,
    destinationState: string;
    destinationCep: string;
    destinationCountry: string;
    distance: string;
    duration: string;
}